Refactor the native face detector
Make i the row index and j the column index. This gives the input
image and the matrices the same layout.

Move weightedSum() into RectFeature and remove the now redundant
scaled* accessors.

+++ NativeFaceDetector/Feature.h
@@ -28,7 +28,7 @@ namespace NativeFaceDetector
 			m_width   = (int)(scale * m_size.width);
 			m_height  = (int)(scale * m_size.height);
 
-			const float area  = (float)(m_width * m_height);
+			const int area    = m_width * m_height;
 			m_invArea         = 1.0f / area;
 			m_scaledThreshold = m_threshold * area;
